Akka Actor Framework Implemented, Redis, Habbo Navigator Handlers
Major updates over the week including:
Akka was added for game logistics using actor-based messaging
Redis was added for session storage, as well as game data caching and
querying
Habbo Navigator handlers and DAO was implemented, almost done with this.

+++ src/main/java/org/mdev/revolution/database/dao/PlayerDao.java
@@ -2,7 +2,7 @@ package org.mdev.revolution.database.dao;
 
 import org.mdev.revolution.database.domain.Player;
 
-public class PlayerDao extends GenericJpaDao<Player, Long> {
+public class PlayerDao extends GenericJpaDao<Player, Integer> {
     public PlayerDao() {
         super();
         super.setClazz(Player.class);
